requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
websockets>=11.0
websocket-client>=1.6.0
aiohttp>=3.9.0
requests>=2.31.0
pyyaml>=6.0
python-binance>=1.0.19
binance-connector>=3.0.0
binance-futures-connector>=1.0.0
pytz>=2023.3

# Scientific computing (for strategy calculations)
scipy>=1.11.0
scikit-learn>=1.3.0

# Data storage
pyarrow>=14.0.0
polars>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# System monitoring
psutil>=5.9.0

# Development
black>=23.0.0
flake8>=6.0.0

# Chart
matplotlib>=3.10.8
